requirements.txt
opencv-contrib-python
pillow
imutils
imagehash
pytesseract
scikit-image